prevent deleting system managed users and roles

+++ services/role.ts
@@ -1,8 +1,14 @@
+import { ApiError } from "js-ts-kit";
 import { Types } from "mongoose";
 import { z } from "zod";
 import {
+	FORBIDDEN_STATUS_CODE,
+	SYSTEM_MANAGED_ROLE_DELETE_ERROR_MESSAGE,
+} from "../constants.ts";
+import {
 	deleteRoleByIdDal,
 	getRoleCountDal,
+	getRoleDal,
 	getRolesDal,
 	upsertRoleDal,
 } from "../dal/role.ts";
@@ -41,6 +47,14 @@ export async function upsertRoleService(
 }
 
 export async function deleteRoleByIdService(_id: Role["_id"]) {
+	const role = await getRoleDal({ _id }).exec();
+	if (role?.systemManaged) {
+		throw new ApiError(
+			SYSTEM_MANAGED_ROLE_DELETE_ERROR_MESSAGE,
+			FORBIDDEN_STATUS_CODE,
+		);
+	}
+
 	const result = await deleteRoleByIdDal(_id).exec();
 	return roleSchema.parse(result);
 }
